blue-ribbon-kennels/brk-site/requirements.txt
amqp==5.0.6
APScheduler==3.8.1
asgiref==3.4.1
autopep8==1.5.7
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
DateTimeRange==1.2.0
Django==3.2.6
django-ckeditor==6.1.0
django-clear-cache==0.3
django-credit-cards==0.4.1
django-crontab==0.7.1
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-nested-admin==3.3.3
django-nested-inline==0.4.4
django-phone-field==1.8.1
django-phonenumber-field==6.0.0
idna==3.3
kombu==5.2.1
mbstrdecoder==1.1.0
numpy==1.21.2
packaging==21.0
pandas==1.3.3
phonenumbers==8.12.37
Pillow==8.3.1
prompt-toolkit==3.0.22
psycopg2==2.9.1
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-monkey-business==1.0.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.26.0
schedule==1.1.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
stripe==2.62.0
toml==0.10.2
typepy==1.3.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
